Optimisations diverses.

+++ libretto/utils.pxd
@@ -17,7 +17,7 @@ cdef list chars_iterator(unicode s)
 cdef ABBREVIATION_RE
 
 
-@cython.locals(out=unicode, i=int, sub=unicode,
+@cython.locals(out=unicode, i=Py_ssize_t, sub=unicode,
                vowels_count=int, vowel_first=bint, j0=int, c0=unicode,
                j1=int, c1=unicode, general_case=bint, particular_case=bint)
 cpdef abbreviate(string, int min_vowels=?, int min_len=?, bint tags=?,
